@@ -3,16 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@reduxjs/toolkit": "^1.9.5",
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.1.5",
+    "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^13.5.0",
+    "custom_dropdown_comp": "^0.1.1",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-data-table-component": "^7.5.3",
-    "react-dom": "^17.0.2",
-    "react-redux": "^8.1.0",
-    "react-router-dom": "^6.10.0",
+    "react-dom": "^18.2.0",
+    "react-focus-lock": "^2.9.4",
+    "react-router-dom": "^6.11.2",
     "react-scripts": "5.0.1",
     "sass": "^1.62.1",
     "web-vitals": "^2.1.4"
@@ -42,7 +42,6 @@
     ]
   },
   "devDependencies": {
-    "better-docs": "^2.7.2",
     "eslint": "^8.42.0",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.32.2",
