@@ -1,14 +1,13 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import ReactDOM from 'react-dom/client'
 import App from './App'
-import { Provider } from 'react-redux'
-import store from './redux/store.js'
+import { EmployeeContextProvider } from './utils/EmployeeContext'
 
-ReactDOM.render(
-  <Provider store={store}>
+const root = ReactDOM.createRoot(document.getElementById('root'))
+root.render(
+  <EmployeeContextProvider>
     <React.StrictMode>
       <App />
     </React.StrictMode>
-  </Provider>,
-  document.getElementById('root')
+  </EmployeeContextProvider>
 )
